+++ pkgs/cupertino_http/example/integration_test/error_test.dart
@@ -3,6 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 
 @Skip('Error tests cannot currently be written. See comments in this file.')
+library;
 
 import 'package:integration_test/integration_test.dart';
 import 'package:test/test.dart';