+++ pkgs/http/lib/src/io_client.dart
@@ -23,7 +23,7 @@ BaseClient createClient() {
 /// Exception thrown when the underlying [HttpClient] throws a
 /// [SocketException].
 ///
-/// Implemenents [SocketException] to avoid breaking existing users of
+/// Implements [SocketException] to avoid breaking existing users of
 /// [IOClient] that may catch that exception.
 class _ClientSocketException extends ClientException
     implements SocketException {