+++ pkgs/http/test/io/request_test.dart
@@ -3,6 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 
 @TestOn('vm')
+library;
 
 import 'package:http/http.dart' as http;
 import 'package:test/test.dart';